sched/walt: Reduce window size to 8 ms and load history to 64 ms
A 20 ms WALT window means that if a vsync happens to start at the same
time as the window, and the CPU frequency is low, the entire frame could
be missed before the schedutil governor responds and ramps up the
frequency in response to the demand. By reducing the load-tracking
window to half of a vsync window, it is much less likely that such an
event will occur. As a result, overall performance is better. This may
also help alleviate the stuttering observed when CPU boosting is disabled.

Bump RAVG_HIST_SIZE_MAX to 8 (64 ms) as well to accomodate this change.
Previously it was 5 (100 ms), but we will have it match the ULPS timeout
used by the display driver which is 64 ms. The total load history
accounted for a task is RAVG_HIST_SIZE_MAX * sched_ravg_window.

Be aware that the 8 ms window is only ideal because the current timer
frequency is 250 Hz, which has a jiffy interval of 4 ms. If the timer
frequency is changed, this will need to be updated for optimal results.
For example, if 300 Hz is in use, the window size should be 10 ms so
that commit d7115a83b36b3 ("sched: walt: Correct WALT window size
initialization") will round it to 9.999999 ms instead of 6.666666 ms.
It may be desirable to adjust RAVG_HIST_SIZE_MAX as well.

If 100 Hz is in use, an 8 ms window will NOT work properly even with the
above initialization commit because integer division rounds down in C.
100 Hz yields a jiffy length of 10 ms, which is not precise enough for
an 8 ms window. The above initialization commit will thus correct the
window size to 0 ms, causing WALT to be disabled and system performance
to be obliterated. The window size should be set to 10 ms when 100 Hz is
the timer frequency.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>